# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小宇宙项目用户交互API、管理后台API",
        "title": "小小の宇宙API文档",
        "contact": {
            "name": "cw4t",
            "email": "cwater4t@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/manager/articles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "获取文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Author ID",
                        "name": "author_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "State",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "CreatedBy",
                        "name": "created_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetArticlesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "添加文章",
                "parameters": [
                    {
                        "description": "文章详情",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddArticleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/manager/articles/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "修改文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Page Title",
                        "name": "page_title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "SEO Title",
                        "name": "seo_title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "SEO URL",
                        "name": "seo_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Related Articles",
                        "name": "related_articles",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Desc",
                        "name": "meta_desc",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Author ID",
                        "name": "author_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Content",
                        "name": "content",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Cover img URL",
                        "name": "cover_image_url",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "State",
                        "name": "state",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "models.ArticleDto": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "author_name": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "category_name": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "cover_image_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "meta_desc": {
                    "type": "string"
                },
                "page_title": {
                    "type": "string"
                },
                "related_articles": {
                    "type": "string"
                },
                "seo_title": {
                    "type": "string"
                },
                "seo_url": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "v1.AddArticleForm": {
            "type": "object",
            "required": [
                "author_id",
                "category_id",
                "content",
                "cover_image_url",
                "language",
                "meta_desc",
                "page_title",
                "related_articles",
                "seo_title",
                "seo_url",
                "state"
            ],
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "cover_image_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string",
                    "default": "jp",
                    "enum": [
                        "jp",
                        "zh",
                        "en"
                    ]
                },
                "meta_desc": {
                    "type": "string"
                },
                "page_title": {
                    "type": "string"
                },
                "related_articles": {
                    "type": "string"
                },
                "seo_title": {
                    "type": "string"
                },
                "seo_url": {
                    "type": "string"
                },
                "state": {
                    "description": "0表示禁用，1表示启用",
                    "type": "integer",
                    "default": 1,
                    "enum": [
                        1,
                        0
                    ]
                }
            }
        },
...
        "v1.GetArticlesResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "lists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ArticleDto"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Token",
            "in": "header"
        }
    }
}